DTOs were organized in a folder
