+++ src/main/java/com/playko/projectManagement/application/handler/IProjectHandler.java
@@ -1,10 +1,8 @@
 package com.playko.projectManagement.application.handler;
 
-import com.playko.projectManagement.application.dto.request.ProjectDeadlineRequestDto;
-import com.playko.projectManagement.application.dto.request.ProjectRequestDto;
+import com.playko.projectManagement.application.dto.request.project.ProjectDeadlineRequestDto;
+import com.playko.projectManagement.application.dto.request.project.ProjectRequestDto;
 import com.playko.projectManagement.application.dto.response.ProjectStatsDto;
-
-import java.time.LocalDate;
 
 public interface IProjectHandler {
     void createProject(ProjectRequestDto projectRequestDto);
